Frames Now Visible
Old main frame is now visible. More refactoring includes removing the "MainPageObjects.py" and "TimerFrameObject.py" due to redundancy and confusing labeling. Nothing removed, just placed into another existing file.
