+++ MainGUI.py
@@ -5,9 +5,7 @@ from tkinter import PhotoImage
 from tkinter import * # Is this necessary?
 from tkinter import ttk # For accessing Notebook widget
 
-import MainPageObjects
 import TimerObject
-import TimerFrameObject
 import TimerTags
 
 ### DETAILS ###
@@ -17,38 +15,54 @@ IMAGES = "ClockResized.gif"
 
 ### OUTERFRAME CLASS  ###
 class OuterFrame:
-    root = tk.Tk() # Creates main root
-    root.title('Our Wonderful Time Tracking GUI 1.0.0') # Title displays at the top
-    #root.geometry = ('500 x 700') # Initial dimensions of window
-    root.minsize(500, 700) # Hmm, this one seems to work and not the root.geometry statement above
+    def __init__(self):
+        self.root = tk.Tk() # Creates main root
+        self.root.title('Our Wonderful Time Tracking GUI 1.0.0') # Title displays at the top
+        #root.geometry = ('500 x 700') # Initial dimensions of window
+        self.root.minsize(500, 700) # Hmm, this one seems to work and not the root.geometry statement above
 
-    ### CREATE NOTEBOOK ###
-    notebook = ttk.Notebook(root) # Create Notebook
+        ### CREATE NOTEBOOK ###
+        self.notebook = ttk.Notebook(self.root) # Create Notebook
 
-    ### CREATE TABS ###
-    mainTab = ttk.Frame(notebook)
-    tagsTab = ttk.Frame(notebook)
-    goalsTab = ttk.Frame(notebook)
+        ### CREATE TABS ###
+        self.mainTab = ttk.Frame(self.notebook)
+        self.tagsTab = ttk.Frame(self.notebook)
+        self.goalsTab = ttk.Frame(self.notebook)
 
-    ### POPULATE NOTEBOOK WITH TABS ###
-    notebook.add(mainTab, text = "Main Menu")
-    notebook.add(tagsTab, text = "Tags Menu")
-    notebook.add(goalsTab, text = "Goal Dashboard")
-    notebook.pack(expand = 1, fill = 'both')
+        ### POPULATE NOTEBOOK WITH TABS ###
+        self.notebook.add(self.mainTab, text = "Main Menu")
+        self.notebook.add(self.tagsTab, text = "Tags Menu")
+        self.notebook.add(self.goalsTab, text = "Goal Dashboard")
+        self.notebook.pack(expand = 1, fill = 'both')
 
-    ### ADD TO mainTab ###
-    Label(mainTab, text = "Put frame with timer, clock image, start/stop button here.").pack()
-    # TODO: Add more widgets here : https://docs.python.org/3/library/tkinter.ttk.html
-    # Add TimerObject, TimerFrameObjects, etc.  Will this be a possible?
+        ### ADD TO mainTab ###
+        #Label(self.mainTab, text = "Put frame with timer, clock image, start/stop button here.").pack()
+        # TODO: Add more widgets here : https://docs.python.org/3/library/tkinter.ttk.html
+        # Add TimerObject, TimerFrameObjects, etc.  Will this be a possible?
+                                                        #^^yes it will >:)
 
-    ### ADD TO tagsTab ###
-    Label(tagsTab, text = "Make new tags here.").pack()
+        ### ADD TO tagsTab ###
+        Label(self.tagsTab, text = "Make new tags here.").pack()
 
-    ### ADD TO goalsTab ###
-    Label(goalsTab, text = "Goals Dashboard here.").pack()
+        ### ADD TO goalsTab ###
+        Label(self.goalsTab, text = "Goals Dashboard here.").pack()
 
-### CREATE OBJECT ###
-outerFrame = OuterFrame()
 
-### MAINLOOP CALL ###
-outerFrame.root.mainloop()
+def main():
+    ### CREATE OBJECT ###
+    outerFrame = OuterFrame()
+
+    #mainTab
+    mainFrame = TimerObject.TimerFrame(outerFrame.mainTab)
+    timer = TimerObject.Timer(outerFrame.mainTab)
+
+    #tagsTab
+    tagFrame = TimerTags.GoalDashboardButton(outerFrame.tagsTab)
+
+    #goalsTab
+    #goalFrame = I dont think this exists yet
+
+    ### MAINLOOP CALL ###
+    outerFrame.root.mainloop()
+
+main()
